moved queue itno its own directory and built labjack w/queue
