+++ driver/ni/BUILD.bazel
@@ -57,7 +57,6 @@ cc_library(
             "error.h",
             "ni.h",
             "scale.h",
-            "ts_queue.h",
             "//client/cpp:synnax",
             "//driver/pipeline",
             "//driver/task",
@@ -68,7 +67,6 @@ cc_library(
             "error.h",
             "ni.h",
             "scale.h",
-            "ts_queue.h",
             "//client/cpp:synnax",
             "//driver/pipeline",
             "//driver/task",
@@ -116,6 +114,7 @@ cc_library(
         "//client/cpp:synnax",
         "//driver/breaker",
         "//driver/config",
+        "//driver/queue",
         "//driver/errors",
         "//driver/loop",
         "//driver/pipeline",
