+++ driver/ni/ni.h
@@ -31,7 +31,7 @@
 
 #include "driver/ni/ai_channels.h"
 #include "driver/ni/error.h"
-#include "driver/ni/ts_queue.h"
+#include "driver/queue/ts_queue.h"
 #include "driver/pipeline/acquisition.h"
 #include "driver/pipeline/control.h"
 #include "driver/task/task.h"
