+++ driver/queue/BUILD.bazel
@@ -1,0 +1,5 @@
+cc_library(
+    name = "queue",
+    hdrs = ["ts_queue.h"],
+    visibility = ["//visibility:public"],
+)
